self_pruning_nn/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.7.0
numpy>=1.24.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
python-multipart>=0.0.9
pillow>=10.2.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
sentence-transformers>=2.6.0
faiss-cpu>=1.8.0
openai>=1.30.0